varios/21/integration/requirements.txt
psycopg2-binary==2.9.9
pytest==8.0.0 